datatypes

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ int main(void)
 	long long int lld;
 	float f;
 
-	printf("Size of a char: %lu byte(s)\n", (unsigned long int)sizeof(c));
-	printf("Size of an int: %lu byte(s)\n", (unsigned long int)sizeof(i));
-	printf("Size of a long int: %lu byte(s)\n", (unsigned long int)sizeof(lu));
-	printf("Size of a long long int: %lu byte(s)\n", (unsigned long int)sizeof(lld));
-	printf("Size of a float: %lu byte(s)\n", (unsigned long int)sizeof(f));
+	printf("Size of a char: %u byte(s)\n", (unsigned int)sizeof(c));
+	printf("Size of an int: %u byte(s)\n", (unsigned int)sizeof(i));
+	printf("Size of a long int: %u byte(s)\n", (unsigned int)sizeof(lu));
+	printf("Size of a long long int: %u byte(s)\n", (unsigned int)sizeof(lld));
+	printf("Size of a float: %u byte(s)\n", (unsigned int)sizeof(f));
 	return (0);
 }
